Title: 앞에서부터 삭제하기 2, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240403/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.cpp
@@ -22,7 +22,7 @@ int main() {
     }
 
     for(int i = 0; i < N - 2; i++){
-        double temp_sum;
+        long long double temp_sum;
         sum -= arr[i];
         if(-arr[i] == pq.top()){
             pq.pop();
